refactor: split robinhood_stock_tools.py into tools/stocks/ subpackage (#67)
Split the 648-line monolithic module into four focused modules:
- stocks/quote.py: price, quote-by-id, pricebook
- stocks/history.py: historical price data
- stocks/fundamentals.py: company info, market hours
- stocks/instruments.py: instrument lookup, search, batch helper

The original robinhood_stock_tools.py is now a thin re-export shim
for backward compatibility. All consumer imports and test patch paths
updated to target the new module locations.
