+++ src/open_stocks_mcp/brokers/robinhood.py
@@ -13,11 +13,11 @@ from open_stocks_mcp.tools.robinhood_account_tools import (
     get_portfolio,
     get_positions,
 )
-from open_stocks_mcp.tools.robinhood_stock_tools import get_stock_price
 from open_stocks_mcp.tools.robinhood_trading_tools import (
     order_buy_market,
     order_sell_market,
 )
+from open_stocks_mcp.tools.stocks.quote import get_stock_price
 
 
 class RobinhoodBroker(BaseBroker):
