+++ src/open_stocks_mcp/tools/stocks/__init__.py
@@ -1,0 +1,28 @@
+"""Stock market data tools subpackage."""
+
+from open_stocks_mcp.tools.stocks.fundamentals import get_market_hours, get_stock_info
+from open_stocks_mcp.tools.stocks.history import get_price_history
+from open_stocks_mcp.tools.stocks.instruments import (
+    _fetch_instruments_batch,
+    find_instrument_data,
+    get_instruments_by_symbols,
+    search_stocks,
+)
+from open_stocks_mcp.tools.stocks.quote import (
+    get_pricebook_by_symbol,
+    get_stock_price,
+    get_stock_quote_by_id,
+)
+
+__all__ = [
+    "_fetch_instruments_batch",
+    "find_instrument_data",
+    "get_instruments_by_symbols",
+    "get_market_hours",
+    "get_price_history",
+    "get_pricebook_by_symbol",
+    "get_stock_info",
+    "get_stock_price",
+    "get_stock_quote_by_id",
+    "search_stocks",
+]
